Trim the Hive distributed query test footprint
The Hive distributed query tests re-run the full AbstractTestJoinQueries, AbstractTestWindowQueries and AbstractTestOrderByQueries suites through the metastore, even though algorithm correctness is already covered by the TPCH-backed distributed runs plus the spill, no-dynamic-filtering and fault-tolerant variants. The only coverage Hive adds is real split sources, dynamic-filter pushdown into the split source, and partitioned and broadcast remote exchanges.

Extract the connector-relevant join shapes into AbstractTestConnectorJoinQueries, which AbstractTestJoinQueries now extends so the engine runs are unchanged, and point the Hive distributed join classes at it. Drop the Hive distributed window and order-by tests entirely: they run no joins, so dynamic filtering does not apply and they add nothing over the TPCH distributed runs while each pays a full query-runner startup.
