+++ plugin/trino-hive/src/test/java/io/trino/plugin/hive/TestHiveDistributedJoinQueriesWithoutDynamicFiltering.java
@@ -14,7 +14,7 @@
 package io.trino.plugin.hive;
 
 import com.google.common.collect.ImmutableMap;
-import io.trino.testing.AbstractTestJoinQueries;
+import io.trino.testing.AbstractTestConnectorJoinQueries;
 import io.trino.testing.QueryRunner;
 import org.junit.jupiter.api.Test;
 
@@ -22,7 +22,7 @@ import org.junit.jupiter.api.Test;
  * @see TestHiveDistributedJoinQueries for tests with dynamic filtering enabled
  */
 public class TestHiveDistributedJoinQueriesWithoutDynamicFiltering
-        extends AbstractTestJoinQueries
+        extends AbstractTestConnectorJoinQueries
 {
     @Override
     protected QueryRunner createQueryRunner()
